refactor: extract 51job manual parsing from resume-normalization.ts (2485→666 lines)
Extract ~1820 lines of 51job-specific manual resume parsing code into a dedicated
resume-normalization.manual51job.ts module. The original file now re-exports from
the new module via `export *`. All consumers import from @trends/shared and are
unaffected — zero behavioral changes.

- resume-normalization.ts: 2485 → 666 lines (–1819 lines, 73% reduction)
- All 2589 tests pass, bun check passes
